requirements.txt
requests==2.32.3
py-flare-common==0.1.6
attrs==24.2.0
python-dotenv==1.0.1
web3==7.11.1
loguru
rich
aioprometheus